Initial import:  * SystemSpecification container class, holding information about the system specs of a device;  * SystemSpecificationLoader that loads the SystemSpecifications.
Changeset:
 * SystemInformation extended with various methods. Can now get total RAM memory, free RAM memory and more;
 * IAgentManager extended and can now give information about the system specs of the agent as well as its performance score;
 * AgentManager unit test extended;
 * Minor refactoring.

[http://jenkins.musala.com/r/6089/ $6089] see #2592

git-svn-id: https://trac.musala.com/ARDC/repo@8891 34708651-a5da-c34d-b20c-10ddab38317d

+++ src/com/musala/atmosphere/commons/sa/DeviceParameters.java
@@ -37,7 +37,7 @@ public class DeviceParameters implements Serializable
 	/**
 	 * Desired available random access memory.
 	 */
-	private int ram;
+	private long ram;
 
 	public int getApiLevel()
 	{
@@ -69,12 +69,12 @@ public class DeviceParameters implements Serializable
 		this.dpi = dpi;
 	}
 
-	public int getRam()
+	public long getRam()
 	{
 		return ram;
 	}
 
-	public void setRam(int ram)
+	public void setRam(long ram)
 	{
 		this.ram = ram;
 	}
